# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LoRa KT API",
    "version": "1.0.0",
    "description": "REST API for team management, checkpoints, RFID access control, LoRa telemetry, and map data."
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Relative to application root"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      },
      "webhookSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Webhook-Secret",
        "description": "Optional server-level secret for the ingest endpoint. When LORA_WEBHOOK_SECRET is set (and not the default placeholder), every POST /api/ingest must include this header."
      }
    },
    "schemas": {
...
      "TeamGroup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "active"
        ]
      },
      "Team": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "number": {
            "type": [
              "integer",
              "null"
            ]
          },
          "organization": {
            "type": [
              "string",
              "null"
            ],
            "description": "Rod / organization the team belongs to"
          },
          "dnf": {
            "type": "boolean",
            "description": "Did not finish; excluded from organization totals."
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamGroup"
            }
          }
        },
        "required": [
          "id",
          "name",
          "groups"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/teams": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "List teams",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "name_asc",
                "name_desc",
                "number_asc",
                "number_desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Teams",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "teams": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Team"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "teams"
        ],
        "summary": "Create team",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "organization": {
                    "type": "string",
                    "description": "Rod / organization the team belongs to"
                  },
                  "dnf": {
                    "type": "boolean",
                    "description": "Did not finish; admin-only."
                  },
                  "number": {
                    "type": "integer"
                  },
                  "group_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "active_group_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created team"
          }
        }
      }
    },
    "/api/teams/{team_id}": {
      "parameters": [
        {
          "name": "team_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "Team detail",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Team",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Team"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "teams"
        ],
        "summary": "Update team",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Team"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated team"
          }
        }
      },
      "delete": {
        "tags": [
          "teams"
        ],
        "summary": "Delete team",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "409": {
            "description": "Conflict"
          }
        }
      },
      "put": {
        "tags": [
          "teams"
        ],
        "summary": "Replace team",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Team"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replaced team"
          }
        }
      }
    },
    "/api/teams/{team_id}/active-group": {
      "post": {
        "tags": [
          "teams"
        ],
        "summary": "Activate a group for the team",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "group_id"
                ],
                "properties": {
                  "group_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Activated"
          }
        }
      }
    },
    "/api/teams/randomize": {
      "post": {
        "tags": [
          "teams"
        ],
        "summary": "Randomize team numbers by group prefix",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "group_id": {
                    "type": "integer",
                    "description": "Optional group id. If omitted, randomize for all groups."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Randomized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "assigned_total": {
                      "type": "integer"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "group_id": {
                            "type": "integer"
                          },
                          "group_name": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "assigned": {
                            "type": "integer"
                          },
                          "needed": {
                            "type": "integer"
                          },
                          "available": {
                            "type": "integer"
                          },
                          "range": {
                            "type": "array",
                            "items": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}